requirements.txt
aioredis==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.25.0
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jose==1.0.0
limits==1.6
passlib==1.7.4
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
redis==3.5.3
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a23
sqlmodel==0.0.6
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
python-dotenv==0.20.0